# refactor(Atom): do not allocate map without atom properties