--- a/include/chemfiles/Atom.hpp
+++ b/include/chemfiles/Atom.hpp
@@ -144,11 +144,12 @@ public:
     /// @example{atom/atomic_number.cpp}
     optional<uint64_t> atomic_number() const;
 
-    /// Get the map of properties associated with this atom. This map might be
-    /// iterated over to list the properties of the atom, or directly accessed.
+    /// Get the map of properties associated with this atom. If no properties
+    /// are set, this function returns `nullopt`. This map might be iterated
+    /// over to list the properties of the atom, or directly accessed.
     ///
     /// @example{atom/properties.cpp}
-    const property_map& properties() const {
+    const optional<property_map>& properties() const {
         return properties_;
     }
 
@@ -158,7 +159,10 @@ public:
     ///
     /// @example{atom/property.cpp}
     void set(std::string name, Property value) {
-        properties_.set(std::move(name), std::move(value));
+        if (!properties_) {
+            properties_ = property_map();
+        }
+        (*properties_).set(std::move(name), std::move(value));
     }
 
     /// Get the `Property` with the given `name` for this atom if it exists.
@@ -168,7 +172,10 @@ public:
     ///
     /// @example{atom/property.cpp}
     optional<const Property&> get(const std::string& name) const {
-        return properties_.get(name);
+        if (properties_) {
+            return (*properties_).get(name);
+        }
+        return nullopt;
     }
 
     /// Get the `Property` with the given `name` for this atom if it exists,
@@ -183,7 +190,10 @@ public:
     /// @example{atom/property.cpp}
     template<Property::Kind kind>
     optional<typename property_metadata<kind>::type> get(const std::string& name) const {
-        return properties_.get<kind>(name);
+        if (properties_) {
+            return (*properties_).get<kind>(name);
+        }
+        return nullopt;
     }
 
 private:
@@ -196,7 +206,7 @@ private:
     /// the atom charge
     double charge_ = 0;
     /// Additional properties of this atom
-    property_map properties_;
+    optional<property_map> properties_ = nullopt;
 
     friend bool operator==(const Atom& lhs, const Atom& rhs);
 };