--- a/src/capi/atom.cpp
+++ b/src/capi/atom.cpp
@@ -188,7 +188,12 @@ extern "C" chfl_status chfl_atom_properties_count(const CHFL_ATOM* const atom, u
     CHECK_POINTER(atom);
     CHECK_POINTER(count);
     CHFL_ERROR_CATCH(
-        *count = static_cast<uint64_t>(atom->properties().size());
+        if (atom->properties()) {
+            *count = static_cast<uint64_t>(atom->properties().value().size());
+        }
+        else {
+            *count = 0;
+        }
     )
 }
 
@@ -197,15 +202,21 @@ extern "C" chfl_status chfl_atom_list_properties(const CHFL_ATOM* const atom, co
     CHECK_POINTER(names);
     CHFL_ERROR_CATCH(
         auto& properties = atom->properties();
-        if (checked_cast(count) != properties.size()) {
+        size_t property_size = 0;
+        if (properties) {
+            property_size = (*properties).size();
+        }
+        if (checked_cast(count) != property_size) {
             set_last_error("wrong data size in function 'chfl_atom_list_properties'.");
             return CHFL_MEMORY_ERROR;
         }
 
-        size_t i = 0;
-        for (auto& it: properties) {
-            names[i] = it.first.c_str();
-            i++;
+        if (properties) {
+            size_t i = 0;
+            for (auto& it: *properties) {
+                names[i] = it.first.c_str();
+                i++;
+            }
         }
     )
 }