@Luthaf 2022-04-15
This is surprising, I would have expected that having an empty map would not allocate/cost us on run-time. I think this is a good solution until we overhaul atomic properties.