@ezavod 2022-04-21
> Sorry, I forgot to merge this!

No worries, I wasn't expecting such a quick response over the holidays :-)

> expected that having an empty map would not allocate/cost us on run-time.

Just FYI: I'm using GCC 7.5.0 (libstdc++). [People on SO](https://stackoverflow.com/a/68561608) also say that there shouldn't be any allocations, so there is probably some other overhead. On first glance there are a lot of constructor/destructor calls for strings caused by the map.